# specs/002-ai-ugc-template-platform/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: AI-UGC Template Platform API
  version: 0.1.0
  description: |
    Public API contract for the transformed AI-UGC template platform.

    Goals:
    - create niche-specific workspaces and project profiles
    - configure LaoZhang or APIMart through one provider abstraction
    - execute configurable AI-UGC generation sessions
    - keep a consistent response envelope independent of provider
servers:
  - url: http://localhost:3000/api
    description: Local development
...
paths:
  /workspaces:
    post:
      tags: [Workspaces]
      summary: Create a workspace template
      operationId: createWorkspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkspaceRequest'
      responses:
        '201':
          description: Workspace created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceResponse'
...
components:
...
  schemas:
    ApiMeta:
      type: object
      required: [timestamp, requestId]
      properties:
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
          format: uuid
...
    WorkspaceTemplate:
      type: object
      required: [id, name, slug, enabledNichePackKeys]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        enabledNichePackKeys:
          type: array
          items:
            type: string
...
    CreateWorkspaceRequest:
      type: object
      required: [name, slug, enabledNichePackKeys]
      properties:
        name:
          type: string
        slug:
          type: string
        enabledNichePackKeys:
          type: array
          items:
            type: string
...
    WorkspaceResponse:
      type: object
      required: [success, data, meta]
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          $ref: '#/components/schemas/WorkspaceTemplate'
        meta:
          $ref: '#/components/schemas/ApiMeta'
...